move custom layer files out of gl/ folders
