+++ bazel/core.bzl
@@ -768,10 +768,10 @@ MLN_OPENGL_SOURCE = [
     "src/mbgl/gl/command_encoder.hpp",
     "src/mbgl/gl/context.cpp",
     "src/mbgl/gl/context.hpp",
-    "src/mbgl/gl/custom_layer.cpp",
-    "src/mbgl/gl/custom_layer_factory.cpp",
-    "src/mbgl/gl/custom_layer_impl.cpp",
-    "src/mbgl/gl/custom_layer_impl.hpp",
+    "src/mbgl/style/layers/custom_layer.cpp",
+    "src/mbgl/layermanager/custom_layer_factory.cpp",
+    "src/mbgl/style/layers/custom_layer_impl.cpp",
+    "src/mbgl/style/layers/custom_layer_impl.hpp",
     "src/mbgl/gl/debugging_extension.cpp",
     "src/mbgl/gl/debugging_extension.hpp",
     "src/mbgl/gl/defines.hpp",
@@ -787,8 +787,8 @@ MLN_OPENGL_SOURCE = [
     "src/mbgl/gl/offscreen_texture.cpp",
     "src/mbgl/gl/offscreen_texture.hpp",
     "src/mbgl/gl/program.hpp",
-    "src/mbgl/gl/render_custom_layer.cpp",
-    "src/mbgl/gl/render_custom_layer.hpp",
+    "src/mbgl/renderer/layers/render_custom_layer.cpp",
+    "src/mbgl/renderer/layers/render_custom_layer.hpp",
     "src/mbgl/gl/render_pass.cpp",
     "src/mbgl/gl/render_pass.hpp",
     "src/mbgl/gl/renderbuffer_resource.hpp",
@@ -822,8 +822,8 @@ MLN_OPENGL_SOURCE = [
 
 MLN_OPENGL_HEADERS = [
     "include/mbgl/gfx/backend.hpp",
-    "include/mbgl/gl/custom_layer.hpp",
-    "include/mbgl/gl/custom_layer_factory.hpp",
+    "include/mbgl/style/layers/custom_layer.hpp",
+    "include/mbgl/layermanager/custom_layer_factory.hpp",
     "include/mbgl/gl/renderable_resource.hpp",
     "include/mbgl/gl/renderer_backend.hpp",
     "include/mbgl/layermanager/location_indicator_layer_factory.hpp",
