+++ platform/default/src/mbgl/layermanager/layer_manager.cpp
@@ -2,9 +2,7 @@
 
 #include <mbgl/layermanager/background_layer_factory.hpp>
 #include <mbgl/layermanager/circle_layer_factory.hpp>
-#ifdef MBGL_RENDER_BACKEND_OPENGL
-#include <mbgl/gl/custom_layer_factory.hpp>
-#endif
+#include <mbgl/layermanager/custom_layer_factory.hpp>
 #include <mbgl/layermanager/fill_extrusion_layer_factory.hpp>
 #include <mbgl/layermanager/fill_layer_factory.hpp>
 #include <mbgl/layermanager/heatmap_layer_factory.hpp>
