+++ platform/ios/platform/darwin/src/MLNOpenGLStyleLayer_Private.h
@@ -2,7 +2,7 @@
 
 #include "MLNStyleLayer_Private.h"
 
-#include <mbgl/gl/custom_layer_factory.hpp>
+#include <mbgl/layermanager/custom_layer_factory.hpp>
 
 namespace mbgl {
