+++ src/mbgl/renderer/layers/render_custom_layer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <mbgl/gl/custom_layer_impl.hpp>
+#include <mbgl/style/layers/custom_layer_impl.hpp>
 #include <mbgl/renderer/render_layer.hpp>
 
 namespace mbgl {
