+++ src/mbgl/style/layers/custom_layer.cpp
@@ -1,6 +1,6 @@
-#include <mbgl/gl/custom_layer.hpp>
-#include <mbgl/gl/custom_layer_impl.hpp>
-#include <mbgl/gl/render_custom_layer.hpp>
+#include <mbgl/style/layers/custom_layer.hpp>
+#include <mbgl/style/layers/custom_layer_impl.hpp>
+#include <mbgl/renderer/layers/render_custom_layer.hpp>
 #include <mbgl/style/layer_observer.hpp>
 
 namespace mbgl {
