+++ src/mbgl/style/layers/custom_layer_impl.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <mbgl/gl/custom_layer.hpp>
+#include <mbgl/style/layers/custom_layer.hpp>
 #include <mbgl/style/layer_impl.hpp>
 #include <mbgl/style/layer_properties.hpp>
